+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * fix copy-to-clipboard on windows, fixes [#344](https://github.com/cars10/elasticvue/issues/344)
 * fix ESC closing modal when search is open in editor, fixes [#346](https://github.com/cars10/elasticvue/issues/346)
+* fix table sorting, fixes [#347](https://github.com/cars10/elasticvue/issues/347)
 
 ## 1.14.0